Replace campaign chart with radar plot
