+++ package.json
@@ -32,6 +32,7 @@
     "react-redux": "^8.0.2",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
+    "recharts": "^3.1.1",
     "web-vitals": "^2.1.4",
     "yup": "^0.32.11"
   },
